+++ client/src/component/Content/Content.js
@@ -3,15 +3,11 @@ import "./content.css";
 import img01 from "../image/watch01.jpg";
 import img02 from "../image/watch02.jpg";
 import img03 from "../image/watch03.jpg";
-import Modal from "../Modal"
-
-
 
 function Content() {
   return (
     <section className="contentSection">
 
-        <Modal/>
 
         <figure className="figure1">
           <img src={img01} className="thumbnail"></img>
