--- client/src/component/Modal/modal.css
@@ -1,34 +1,0 @@
-.btn{
-    position: absolute;
-    color: blue;
-    font-size: 20px;
-    font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif;
-}
-
-.modal{
-    position: absolute;
-    background-color: white;
-    border:1px solid black;
-    height: 150px;
-    width: 400px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-input{
-    width: 100%;
-    margin:5px;
-    border:none;
-    background-color: black;
-    color:white;
-    font-size: 15px;
-    text-align: center;
-}
-
-button{
-    width: 100%;
-    border: none;
-    background-color: black;
-    color: white;
-}
