+++ client/src/component/Navigation/nav.css
@@ -1,3 +1,5 @@
+
+
 .nav-container{
     background-color: turquoise;
     position: absolute;
@@ -17,7 +19,50 @@ ul{
 }
 
 li:hover{
-    color: blue;
+    cursor: pointer;
     scale: 1.25;
     transition: 1s;
+}
+
+
+.modalForm{
+    border-radius: 10%;
+    position: absolute;
+    background-color: white;
+    border:5px solid black;
+    height:400px;
+    width:350px;
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+    align-items: center;
+    z-index: 1;
+    left:37vw;
+    top:25vh;
+}
+
+
+input{
+    width: 200px;
+    height: 30px;
+    border-style:none;
+    background-color: black;
+    text-align: center;
+    color: white;
+    font-family: Arial, Helvetica, sans-serif;
+}
+
+button{
+    width: 204px;
+    height: 30px;
+    border: none;
+    background-color: black;
+    color: white;
+    font-family: Arial, Helvetica, sans-serif;
+}
+
+button:hover{
+    scale:1.025;
+    transition: 1s;
+    
 }
